requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
edge-tts==7.0.2
frozenlist==1.7.0
idna==3.10
keyboard==0.13.5
Levenshtein==0.27.1
MouseInfo==0.1.3
multidict==6.4.4
numpy==2.3.0
packaging==25.0
pillow==11.2.1
propcache==0.3.2
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
RapidFuzz==3.13.0
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.14.0
yarl==1.20.1